Version 1

+++ test.c
@@ -8,9 +8,9 @@ int main()
 {
    char item1[50];
    char item2[50];
-   int price1, price2;
+   float price1, price2;
    int qty1, qty2;
-   int total1, total2, totalCost;
+   float total1, total2, totalCost;
    int question;
 
    // First item
@@ -22,7 +22,7 @@ int main()
    scanf("%d", &qty1);
 
    printf("\nWhat is the price of %s: ", item1);
-   scanf("%d", &price1);
+   scanf("%f", &price1);
 
    // Second item
    printf("\nDo you want to buy something else? (If yes select <1> , If no select <2>) =>");
@@ -36,14 +36,14 @@ int main()
    scanf("%d", &qty2);
 
    printf("\nWhat is the price of %s: ", item2);
-   scanf("%d", &price2);
+   scanf("%f", &price2);
 
    // Calculations
    total1 = qty1 * price1;
    total2 = qty2 * price2;
    totalCost = total1 + total2;
 
-   printf("\nThe total price = %d\n", totalCost);
+   printf("\nThe total price = %f\n", totalCost);
 
    // Payment Section
    printf("\nSelect payment method (Verve or Visa): ");
@@ -60,7 +60,7 @@ int main()
  // Calculations
    total1 = qty1 * price1;
 
-   printf("\nThe total price = %d\n", total1);
+   printf("\nThe total price = %f\n", total1);
 
    // Payment Section
    printf("\nSelect payment method (Verve or Visa): ");
@@ -83,6 +83,8 @@ int main()
      return 0;
 }
    
+   
+
